subindo testes e mais umas funções

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: dsayumi- <dsayumi-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 21:45:01 by dsayumi-          #+#    #+#             */
-/*   Updated: 2023/05/10 23:17:25 by dsayumi-         ###   ########.fr       */
+/*   Updated: 2023/05/13 18:36:58 by dsayumi-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	size_t	len;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	len = ft_strlen(src);
